requirements.txt
Flask==2.3.3
newspaper3k==0.2.8
beautifulsoup4==4.12.2
requests==2.31.0
gunicorn==21.2.0
